#1057: Factor out search form into separate component for inclusion on home page
