+++ src/app/app.module.ts
@@ -74,7 +74,6 @@ import { ContactService } from '@services/contact.service';
 import { AboutComponent } from '@about/about.component';
 import { ConfirmComponent } from '@confirm/confirm.component';
 import { CreateContactService } from '@app/create-contact/create-contact.service';
-import { SearchDialogService } from '@app/search-dialog/search-dialog.service';
 import { SearchService } from '@services/search.service';
 import { SavedSearchesComponent } from '@saved-searches/saved-searches.component';
 import { DatePipe } from '@angular/common';
@@ -132,6 +131,8 @@ import { PapaParseModule } from 'ngx-papaparse';
 import { EventPublicReportComponent } from '@event-public-report/event-public-report.component';
 import { SearchResultsSummaryReportComponent } from '@search-results-summary-report/search-results-summary-report.component';
 import { ViewNotificationDetailsComponent } from '@notifications/view-notification-details/view-notification-details.component';
+import { SearchFormComponent } from './search-form/search-form.component';
+import { SearchFormService } from './search-form/search-form.service';
 
 
 @NgModule({
@@ -187,6 +188,7 @@ import { ViewNotificationDetailsComponent } from '@notifications/view-notificati
     EventPublicReportComponent,
     SearchResultsSummaryReportComponent,
     ViewNotificationDetailsComponent,
+    SearchFormComponent,
   ],
   imports: [
     BrowserModule,
@@ -262,7 +264,7 @@ import { ViewNotificationDetailsComponent } from '@notifications/view-notificati
     OrganizationService,
     RoleService,
     SearchService,
-    SearchDialogService,
+    SearchFormService,
     UserService,
     LocationSpeciesDiagnosisService,
     SpeciesDiagnosisService,
