+++ src/app/events/events.component.ts
@@ -14,7 +14,6 @@ import { EventService } from '@services/event.service';
 import { Router, ActivatedRoute } from '@angular/router';
 
 import { SearchDialogComponent } from '@search-dialog/search-dialog.component';
-import { SearchDialogService } from '@app/search-dialog/search-dialog.service';
 
 import { DisplayQuery } from '@interfaces/display-query';
 
@@ -29,6 +28,7 @@ import { ConfirmComponent } from '@confirm/confirm.component';
 import { EventGroupManagementComponent } from '@app/event-group-management/event-group-management.component';
 import { EventGroupManagementService } from '@services/event-group-management.service';
 import { CurrentUserService } from '@services/current-user.service';
+import { SearchFormService } from '@app/search-form/search-form.service';
 
 
 @Component({
@@ -96,7 +96,7 @@ export class EventsComponent implements AfterViewInit, OnInit {
     public snackBar: MatSnackBar,
     private eventService: EventService,
     private resultsCountService: ResultsCountService,
-    private searchDialogService: SearchDialogService,
+    private searchFormService: SearchFormService,
     private currentUserService: CurrentUserService,
     private eventGroupManagementService: EventGroupManagementService,
     private router: Router,
@@ -112,7 +112,7 @@ export class EventsComponent implements AfterViewInit, OnInit {
       this.eventCount = count;
     });
 
-    this.searchQuerySubscription = this.searchDialogService.getSearchQuery().subscribe(
+    this.searchQuerySubscription = this.searchFormService.getSearchQuery().subscribe(
       searchQuery => {
 
         // this.resultsLoading = true;
@@ -151,7 +151,7 @@ export class EventsComponent implements AfterViewInit, OnInit {
         this.selectedEventGroup = selectedEventGroup;
       });
 
-    this.searchQuerySubscription = this.searchDialogService.getDisplayQuery().subscribe(
+    this.searchQuerySubscription = this.searchFormService.getDisplayQuery().subscribe(
       displayQuery => {
         this.currentDisplayQuery = displayQuery;
         console.log('New display query: ' + this.currentDisplayQuery);
